+++ app/routes/project.js
@@ -1,8 +1,9 @@
 import Route from '@ember/routing/route';
-import ScrollTracker from 'ember-api-docs/mixins/scroll-tracker';
 import { inject as service } from '@ember/service';
+import { withScrollReset } from '../utils/with-scroll-reset';
 
-export default class ProjectRoute extends Route.extend(ScrollTracker) {
+@withScrollReset
+export default class ProjectRoute extends Route {
   /** @type {import('@ember/routing/router-service').default} */
   @service
   router;
